♻️ Refactor starport to use AST analysis to modify protobuf and golang
This replaces the use of placeholders and regex modification. It also
adds two new packages `gocode` and `protocode` which are used to query,
build, and insert AST constructs into the 'on file' disk.

The gocode package is much heavier and more complex (and thus, a bit
messier) than the protocode package. Sadly, the code is also not *less*
than the code generated by strings. However we gain the ability to
validate the golang AST before writing to a file, so we can know what
went wrong (to a point, as the golang compiler is used and its error
messages make a 3 year old look like shakespeare and homer all wrapped
into one)

🎨 Run starport through gofumpt (instead of gofmt)

Signed-off-by: Izzy Muerte <63051+bruxisma@users.noreply.github.com>