♻️ Further simplify operations in templates/typed/list/genesis.go
More work is of course needed, but we're very close to a "simple"
approach to implementing these various AST mutations in a moderately
simple way. This could possibly even be scripted in the future, and not
require code to be compiled in to starport.

⚡️ Improve file write performance

For performance reasons we no longer need to generate a string *before*
writing the file out, and this saves us both memory and CPU time.